requirements.txt
pickleDB==0.9.2
python-telegram-bot==13.7
requests